final/parking_app/requirements.txt
aes==1.2.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.6
Django==4.2.13
django-cors-headers==3.14.0
django-filter==24.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
dramatiq==1.16.0
fastapi==0.110.3
filelock==3.10.5
gmpy2==2.1.5
greenlet==3.0.3
h11==0.14.0
idna==3.6
Mako==1.3.3
MarkupSafe==2.1.5
mpmath==1.3.0
platformdirs==3.2.0
prometheus_client==0.20.0
psycopg2-binary==2.9.9
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==1.7.1
pytz==2023.3
redis==5.0.4
requests==2.31.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.4.3
starlette==0.37.2
sympy==1.12
typing_extensions==4.11.0
tzdata==2023.2
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.21.0
